Refactoring, quality and validation

+++ src/ChatNotifyService/ChatNotifyService.API/Controllers/ChatController.cs
@@ -1,9 +1,9 @@
-using System.Security.Claims;
-using ChatNotifyService.ABS.IEntities;
+using ChatNotifyService.ABS.Dtos;
+using ChatNotifyService.ABS.Models;
 using ChatNotifyService.ABS.IHelpers;
 using ChatNotifyService.ABS.IServices;
-using ChatNotifyService.BLL.Dtos.Create;
-using ChatNotifyService.BLL.Dtos.Send;
+using ChatNotifyService.API.Helpers;
+using ChatNotifyService.BLL.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -18,11 +18,9 @@ namespace ChatNotifyService.API.Controllers;
 [ApiController]
 public class ChatController(
     IChatService chatService,
-    IHttpContextAccessor httpContextAccessor,
-    IBigMapper<IChat, ChatDto, ChatDtoShort> chatMapper,
-    ICreateMapper<IChat, ChatCreateDto> chatCreateMapper,
-    IBigMapper<IChatMember, MemberDto, MemberDtoShort> memberMapper,
-    ICreateMapper<IChatMember, MemberCreateDto> memberCreateMapper)
+    GetUserHelper getUserHelper,
+    IBigMapper<Chat, ChatDto, ChatDtoShort> chatMapper,
+    IBigMapper<ChatMember, MemberDto, MemberDtoShort> memberMapper)
         : ControllerBase
 {
     /// <summary>
@@ -32,13 +30,8 @@ public class ChatController(
     [HttpGet]
     public async Task<ActionResult<IEnumerable<ChatDtoShort>>> GetAllChats([FromRoute] Guid spaceId)
     {
-        var senderIdClaim = httpContextAccessor.HttpContext?
-            .User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-            
-        if (senderIdClaim == null)
-            return Unauthorized("Invalid token.");
-            
-        var memberId = Guid.Parse(senderIdClaim);
+        Guard.AgainstEmptyGuid(spaceId);
+        var memberId = getUserHelper.GetCurrentUserId();
         var chats = await chatService.GetAllChatsAsync(spaceId, memberId);
         return Ok(chats.Select(chatMapper.ToShortDto));
     }
@@ -50,12 +43,11 @@ public class ChatController(
     [HttpGet("{chatId:guid}")]
     public async Task<ActionResult<ChatDto>> GetChatById([FromRoute] Guid chatId)
     {
-        var memberId = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
+        Guard.AgainstEmptyGuid(chatId);
+        var memberId = getUserHelper.GetCurrentUserId();
+
         var chat = await chatService.GetChatByIdAsync(chatId, memberId);
-        if (chat == null)
-            return NotFound();
-        
-        return Ok(chatMapper.ToDto(chat));
+        return chat == null ? NotFound() : Ok(chatMapper.ToDto(chat));
     }
 
     /// <summary>
@@ -67,13 +59,7 @@ public class ChatController(
     public async Task<ActionResult<ChatDto>> CreateChat(
         [FromRoute] Guid spaceId, [FromBody] ChatCreateDto chat)
     {
-        var senderIdClaim = httpContextAccessor.HttpContext?
-            .User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-        
-        if (senderIdClaim == null)
-            return Unauthorized("Invalid token.");
-        
-        var senderGuid = Guid.Parse(senderIdClaim);
+        var senderGuid = getUserHelper.GetCurrentUserId();
         if (chat.Members.All(m => m.MemberId != senderGuid))
         {
             chat.Members.Add(new MemberCreateDto
@@ -88,10 +74,9 @@ public class ChatController(
             senderMember.PermissionLevel = PermissionLevel.Admin;
         }
         
-        var newChat = chatCreateMapper.ToEntity(spaceId, chat);
-        var createdChat = await chatService.CreateChatAsync(newChat);
+        var createdChat = await chatService.CreateChatAsync(chat);
         return Created(
-            "chat/" + createdChat.Id,
+            new Uri("chat/" + createdChat.Id),
             chatMapper.ToDto(createdChat));
     }
     
@@ -102,15 +87,15 @@ public class ChatController(
     [HttpPut]
     public async Task<ActionResult<ChatDto?>> UpdateChat([FromBody] ChatDtoShort updatedChat)
     {
-        var memberId = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
-        //if (!await chatService.IsMemberAsync(updatedChat.Id, memberId))
-        //    return Forbid("You are not a member of this chat.");
+        Guard.AgainstNull(updatedChat);
+        var memberId = getUserHelper.GetCurrentUserId();
+        if(!await getUserHelper.CheckMemberPermissionInChatAsync(updatedChat.Id, PermissionLevel.Admin))
+        {
+            return Forbid();
+        }
 
         var chat = await chatService.UpdateChatAsync(chatMapper.ToEntity(updatedChat));
-        if (chat == null)
-            return NotFound();
-        
-        return Ok(chatMapper.ToDto(chat));
+        return chat == null ? NotFound() : Ok(chatMapper.ToDto(chat));
     }
     
     /// <summary>
@@ -120,12 +105,14 @@ public class ChatController(
     [HttpDelete("{chatId:guid}")]
     public async Task<IActionResult> DeleteChat([FromRoute] Guid chatId)
     {
-        var memberId = Guid.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
+        Guard.AgainstEmptyGuid(chatId);
+        var memberId = getUserHelper.GetCurrentUserId();
+        if(!await getUserHelper.CheckMemberPermissionInChatAsync(chatId, PermissionLevel.Admin))
+        {
+            return Forbid();
+        }
         var deleted = await chatService.DeleteChatAsync(chatId, memberId);
-        if (!deleted)
-            return NotFound();
-        
-        return NoContent();
+        return deleted ? NoContent() : NotFound();
     }
     
     /// <summary>
@@ -137,6 +124,8 @@ public class ChatController(
     public async Task<ActionResult<IEnumerable<MemberDto>>> GetChatMembers(
         [FromRoute] Guid spaceId, [FromRoute] Guid chatId)
     {
+        Guard.AgainstEmptyGuid(spaceId);
+        Guard.AgainstEmptyGuid(chatId);
         var members = await chatService.GetChatMembersAsync(spaceId, chatId);
         return Ok(members.Select(memberMapper.ToDto));
     }
@@ -148,9 +137,15 @@ public class ChatController(
     /// <param name="newMember"></param>
     [HttpPost("{chatId:guid}/members")]
     public async Task<ActionResult<MemberDto?>> AddMemberToChat(
-        [FromRoute] Guid chatId, [FromBody] MemberCreateDto newMember)
+        [FromRoute] Guid chatId, [FromBody] MemberDtoShort newMember)
     {
-        var member = await chatService.AddMemberToChatAsync(memberCreateMapper.ToEntity(chatId, newMember));
+        Guard.AgainstEmptyGuid(chatId);
+        Guard.AgainstEmptyGuid(newMember.MemberId);
+        if (!await getUserHelper.CheckMemberPermissionInChatAsync(chatId, PermissionLevel.Admin))
+        {
+            return Forbid();
+        }
+        var member = await chatService.AddMemberToChatAsync(newMember);
         return Ok(memberMapper.ToDto(member));
     }
     
@@ -163,10 +158,14 @@ public class ChatController(
     public async Task<IActionResult> RemoveMemberFromChat(
         [FromRoute] Guid chatId, [FromRoute] Guid memberId)
     {
-        var removed = await chatService.RemoveMemberFromChatAsync(chatId, memberId);
-        if (!removed)
-            return NotFound();
+        Guard.AgainstEmptyGuid(chatId);
+        Guard.AgainstEmptyGuid(memberId);
         
-        return NoContent();
+        if (!await getUserHelper.CheckMemberPermissionInChatAsync(chatId, PermissionLevel.Admin))
+        {
+            return Forbid();
+        }
+        var removed = await chatService.RemoveMemberFromChatAsync(chatId, memberId);
+        return removed ? NoContent() : NotFound();
     }
 }
